do operation and add date time details in api response.

+++ src/main/java/com/project/mausam/provider/openweather/dto/CityOpenWeatherResponse.java
@@ -39,7 +39,7 @@ public class CityOpenWeatherResponse {
 	private Clouds clouds;
 	private Long dt;
 	private Sys sys;
-	private Long timezone;
+	private Integer timezone;
 	private Long id;
 	private String name;
 	private Long cod;
